revert r1171729, as those changes broke compilation on platforms where qreal != double
svn path=/trunk/KDE/kdelibs/; revision=1196092

+++ kio/kfile/kimagefilepreview.cpp
@@ -235,8 +235,8 @@ void KImageFilePreview::KImageFilePreviewPrivate::_k_slotStepAnimation( int fram
 
     imageLabel->setPixmap(pm);
 
-    m_pmCurrentOpacity = qMax(m_pmCurrentOpacity - 0.4, qreal(0.0));
-    m_pmTransitionOpacity = qMin(m_pmTransitionOpacity + 0.4, qreal(1.0));
+    m_pmCurrentOpacity = qMax(m_pmCurrentOpacity - 0.4, 0.0);
+    m_pmTransitionOpacity = qMin(m_pmTransitionOpacity + 0.4, 1.0);
 }
 
 void KImageFilePreview::KImageFilePreviewPrivate::_k_slotFinished()
